CONFIG_DOM_HTML - Introduce stub for experimental DOM-based HTML parser
Currently living in document/html/dom.*, this aims to be an alternative
for the original Mikulas' HTML parser, using the experimental DOM engine.
It does not aim to replace Mikulas' HTML renderer just yet.

It is shifted into the existing infrastructure pretty uglily now, it will need
more hacking to make this a bit cleaner. But surprisingly, only three files
outside of document/html needed ifdefs, two of them in document/. Nifty!