Add support for reformatting lines in source high-lighting
If wrapping has been enabled, reformat and indent source lines so
start and end tag are on the same column unless only one element
is on the line in which case tags are kept on a single line.

It is slightly abuse of the wrap option but it could be turned into
even wrapping the text nodes.