Move debian to contrib/debian.
Today on the elinks-dev mailing list, the new maintainer of the Debian
elinks package asked the debian directory to be removed from released
tarballs, because it makes the Debian diff hard to read.  Also, the
files there are out of date.

When the debian directory was discussed in April 2007, people didn't
want it to be removed, so I instead moved it to contrib/debian in
ELinks 0.12.GIT (599cf3d91e16b083835a9c5eec2cad623bb6a66c).
And now I'm doing the same in 0.11.3.GIT.