Remove Cogito from ChangeLog and INSTALL too
Git does not appear to have anything like cg clone -s, which makes the
clone in the current directory rather than in a new directory.  It
seems possible to work around that with:

git clone --bare http://elinks.cz/elinks.git .git
git --git-dir=.git config core.bare false
git reset --hard
git clean -f
git checkout elinks-0.12

but that's already so complex that I think it'll be easier to just
remove the whole directory and clone to a new one.

The whole concept of first downloading a snapshot and then updating
that from version control comes from the time when ELinks was in CVS.
It made sense then because CVS could download deltas based on the data
in the CVS subdirectories contained in the snapshots.  However, Git
wants to get the whole history and does not benefit from having the
files of a single commit available in advance.