Bug 765: use load_uri to load old tab's document
Kalle reported that after commit 5c96d430c98d7b67cf397aac83cfdfa1788aa6df,
ELinks would crash if the document in the old tab was still loading when a
new tab was opened.  The problem was that the new session's download.data
pointer was not updated to point to the session as doc_loading_callback
expects.

Instead of just calling render_document_frames, set up the download and
call load_uri.