Don't follow wrong link when an event handler changes documents under us
When a link had an onClick event handler that changes the current document,
ELinks would follow the current link of the document displayed after
executing the handler.

Factor goto_link out of goto_current_link.

Use goto_link instead of goto_current_link in activate_link to ensure that
the link that is passed in by enter() is followed.