AUTHORS: Add a licensing statement.
It is all in one line because <http://elinks.cz/authors.html>
is converted from AUTHORS and the conversion script is not
clever enough to know which newlines should be kept.

If the licence of ELinks is ever changed to "GNU GPL v3 or later",
I'll probably stop hacking it, but I don't want to legally forbid
such a change.